MB-27860 Hash Join Support for ANSI JOIN
This checkin adds support for Hash Join to handle ANSI JOIN queries.

Hash join is considered when a new USE HASH join hint is specified. There are
two versions of the new hint: USE HASH(build) or USE HASH(probe), to indicate
whether the keyspace is to be used as build side or probe side of a hash join.
When USE HASH hint is specified, the planner considers hash join for the
current join, and if successful, hash join will be used. If not successful,
then the planner considers nested-loop join as it does today.

The new HashJoin operator contains a child operator which is the build side.
At runtime the hash table is built first (in beforeItems()) and the probe
side is subsequently processed.

For outer join the subservient side must be the build side of the hash join.
For ANSI NEST the right-hand side also must be the build side of the hash join.

Change-Id: Ic2099c6024587d4a71514d006f5fb9adf71a70d8
Reviewed-on: http://review.couchbase.org/88952
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>