MB-27860 Duplicate values should not contribute to hash table load factor
The load factor is used to determine when a hash table should double in size,
it should be affected by the number of occupied buckets in the hash table, and
thus when we have duplicated hash values (which occupy the same bucket) the
duplicates should not contribute to the load factor.

Introduce a distinct value in the hash table, which should corresponds to the
number of occupied buckets, which is what should be used for determining the
load factor.

Change-Id: I5803f0c3b22ec763b37eb5304675e51ac23088cd
Reviewed-on: http://review.couchbase.org/89122
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>