MB-28018 Hash Join returns errors when joining on meta id
With the new hash join support (MB-27860) if the join condition is on meta id
of the right-hand-side keyspace, then it's possible that no viable plan is
generated. Current code does not detect this properly and generates an empty
scan which leads to runtime issues.

The fix is two-fold. First, allow a primary scan on document key if hash join
is being considered. Second, adds a check that if no viable plan is generated,
bail out instead of generating a hash join plan with an empty scan.

Change-Id: I1a3bb0f1a684064bc4588a10b235f1d29df12100
Reviewed-on: http://review.couchbase.org/89193
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>