MB-27860 Properly formalize USE KEYS expression in ANSI JOIN
Previously USE KEYS is not allowed on the right-hand side keyspace for an
ANSI JOIN, thus the code for formalizing USE KEYS expression is inside
the non-ANSI JOIN block. Since we now allow USE KEYS for the right-hand
side keyspace for ANSI JOIN, we need to move the code for formalizing
USE KEYS expression out of the block for non-ANSI JOIN, otherwise
identifiers inside a subquery under USE KEYS will not be properly
formalized.

Change-Id: Ifd8d8144d0ba52e4301f1271b92b2da4e6f8b846
Reviewed-on: http://review.couchbase.org/89345
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>