MB-27860 Fix runtime error in Hash Join when parallelism is on
The Hash Join operator is not parallelized, however the operator above
Hash Join could be parallel. In such case when RunOnce() is called
multiple times on Hash Join, only one of them will actually execute,
however there currently is a defer function to drop the hash table
which gets triggered if the hash join code is NOT being executed,
which results in hash table being dropped while hash join is
executing (in another thread). This causes runtime issues.

The fix is to remove the defer function call and rely on the hash table
drop function in afterItems() to handle the drop.

Change-Id: I7652867c83e1547e14f052fab78f63c7cad444bf
Reviewed-on: http://review.couchbase.org/89358
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>