MB-27028 ANSI JOIN query may miss rows in rare cases
When nested-loop join is used for ANSI JOIN, the inner side of the join
needs to be reopened for each outer document. Just before reopen a
SendStop() call is made to the inner (child) operator. In majority of
cases when the child operator is closed already the SendStop() call is
a no-op, however in rare cases a stop signal is sent to the stop channel
of the underlying index scan (before the index scan is completely closed).
In such cases the stop signal may end up stopping the next iteration of
the index scan execution, causing missed rows from the join.

The fix is to drain the stop channel before next execution.

Change-Id: I8dbefd9960731f44ebbdbc05f508463b8ceb76c7
Reviewed-on: http://review.couchbase.org/89522
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>