MB-27860 Make USE HASH hint work with RIGHT OUTER JOIN
Join hint (such as USE HASH) cannot be specified on the first keyspace
in a query. If a RIGHT OUTER JOIN is being performed, then a USE HASH
hint can be specified on the right-hand side of the join, which is the
dominant side for the outer join. Since in planner we only check for
join hint on the subservient side, we need to transfer the join hint
from the dominant side to the subservient side in case of RIGHT OUTER
JOIN.

Change-Id: If69752b30b26ce49253a116a930d84d02c800560
Reviewed-on: http://review.couchbase.org/90004
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>