MB-28362 Reset internal value for expression when expression changes
ExpressionBase has a value field, and it's populated the first time
Value() function is called on the expr. When the expression has
been changed, this value field is not reset thus next time Value()
is called it may return the wrong result.

Properly handle the case for expression.replacer, whenever an
expression is replaced with another expression, reset the value field
so it needs to be reevaluated.

Change-Id: I1b41317a18770cb5c77c92e6564e1f80f63f0ba7
Reviewed-on: http://review.couchbase.org/90019
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>