MB-28533 Properly formalize keyspace reference inside Meta() functions
This is a regression from fix for MB-27706, where an identifier expression now
has a marking to indicate whether the identifier is a keyspace reference. Then
in Formalizer.VisitIdentifer() function this information is used to determine
whether a keyspace reference needs to be added in front of an identifier. The
issue with this defect is that when a meta() function adds a keyspace reference
it did not properly set the indicator, resulting in another keyspace reference
added in front of it.

The fix is to properly mark the new identifier as a keyspace reference. This is
done in a couple of places where new identifiers are generated.

Also a few code cleanup is done, including:
  - check this.indexScope() instead of this.mapSelf() for keyspace references
  - add keyspace marking to identifiers when it's a user-specified reference

Change-Id: Ie19d3463787be06ae6016fa9a3c8225bb6452761
Reviewed-on: http://review.couchbase.org/90612
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>