MB-28720 Include certain join filters when considering UNNEST SCAN
Unnest scan is special in that it handles two keyspace references at once,
i.e., if you do keyspace1 UNNEST keyspace1.array as keyspace2, and there
exists an array index on keyspace1.array, then we can generate an UNNEST
SCAN utilizing the array index. The issue is that when probing the index
we need to include any predicates that references both keyspace references,
not just predicates referring to a single keyspace like we usually do
(not counting the keyspaces that we've already done planning).

Add logic to detect whether a join filter ONLY references the two relevant
keyspace references, and if so, include such join filter when probing the
array index.

Change-Id: I703c7224584d18a3639ce050df26efe5e9a86b2c
Reviewed-on: http://review.couchbase.org/91048
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>