MB-27886 intersect ordered scan ditches filters documents in error
I have finally tracked it down to the child signal being asynchronous compared to the document keys
in the value exchange subsystem, meaning that when we see that the first scan finish, we still have
lots of keys from lots of children whose bits have to be added to the bits in the key queue.
This is different from when we where using channels, because the child signal there would arrive
after all the keys from that child had been processed.

Essentially, we have to process all the existing keys using the full bits until all keys have
been exhausted, and switch to the child bits only for the keys that remain in the queue.
This is actually simpler code than we had for channels.

Change-Id: I52b13580940e9a5d1368973d372264c28886e000
Reviewed-on: http://review.couchbase.org/91108
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>