MB-28605 OR expression does not evaluate to NULL if one subterm evaluates to NULL
The problem is the Value() call on an expression, where it loops through all children
and if one child evaluates to MISSING or NULL, then the entire expression evaluates
to MISSING or NULL. This is ok for most cases including AND terms, but for OR, it is
not right. NULL or <valid_expr> should evaluates to <valid_expr>, not NULL. Similar
for MISSING.

The fix is to add a flag to expression to indicate whether the Value() call evaluates
to MISSING or NULL. This flag is set inside the Value() call. Then:
  - add a Value() call to expression.Or to override the default, in there we need to
    detect and skip subterms which evaluates to MISSING or NULL and consider the rest
    of the subterms. If no other subterms then MISSING or NULL is returned, otherwise
    go with the other subterms.
  - when classifying expression, enhance the VisitOr() function to loop through
    subterms of OR, and skip (throw away) subterms that evaluates to MISSING or NULL,
    this is necessary since we don't want to use such expressions to incorrectly
    generate an index span, for example.
  - no need to do similar processing for VisitAnd() since if one of the AND subterm
    evaluates to MISSING or NULL, then the entire clause evaluates to MISSING or NULL,
    and current code already handle that situation (empty plan generated).

There are a large number of files changed in expression package, but majority of them
are simple replacements since I replaced a few boolean variables with flags.
In addition, the Function interface previously defined a function Volatile() which
does not seem to be used anywhere so I removed it, to avoid potential confusion with
the newly introduced volatile() function on the base.

Change-Id: I9c3cf24efea12cfe59a9d01791b9a5bcdcc1ed23
Reviewed-on: http://review.couchbase.org/91482
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>