MB-28882 Properly handle constant filters (TRUE or FALSE) in planner
Although not a common situation, constant filters do appear in queries,
e.g., select * from default where 1 = 1; or sometimes filters can be
evaluated to constant such as "c1 = null" which evaluates to NULL which
is handled like FALSE (in WHERE clause). Previouly since the entire
WHERE clause is always used in the planner such situations are handled,
however with the newly introduced expression handling framework in
vulcan, we only attach filters that references 1 or more keyspaces,
which means constant filters are not always properly handled.

Introduce a new "constant" holder when classifying expressions such that
we keep track of whether TRUE or FALSE constant are used as filters, and
if so, when the caller is handling WHERE clause, set flag to indicate
the WHERE clause is TRUE or FALSE, and put logic in the planner to
properly handle such situations.

Change-Id: I4f31735adeada1c594d7775d764cae674e262015
Reviewed-on: http://review.couchbase.org/91536
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>