MB-29003 add options to NewParsedValue
Currently parsed values cache scanned fields only if they are deemed large enough for the caching to be meaningful.
This may work for N1QL, but not for the projector.
Now that the projector uses the same parsed value for all the WHERE clauses and all the projections,
it makes sense to give the projection the option of deciding when it wants caching.

John - just change the current call to NewParsedValue() to NewParsedValueWithOptions(), and then you have control as to when you want to cache scanned fields.
I suspect that turning on caching at all times will work for you, but probably you need to do some testing first.

Change-Id: I7a8f0e8fea83b2ca2c87dd7db9e6dd050bda5011
Reviewed-on: http://review.couchbase.org/92041
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>