MB-28970 n1ql panics on indexer error
The issue that we have here is that a stop could be sent asynchronously while a request has already terminated as is going through the cleanup phase.
While we need to propagate the stop to all goroutines even if we are in cleanup (if all goroutines don't get stopped, cleanup won't complete),
we have to be mindful that some operators may have already cleaned up after themselves: skip all those children that have already been nilified.

Change-Id: I3cab4bcff7e7606d470ca9d687dab75279ac0659
Reviewed-on: http://review.couchbase.org/92052
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>