MB-28662 Array index and OR clause may not cover the query properly
With array index key we include ANY clause as filter cover. If there is an OR
clause, and the same ANY clause appears on both arms of the OR, current code
does not properly add the ANY clause as filter cover since current code uses
a map with expression pointer as key of the map, and although two expressions
may be the same, they appear different since they have different expression
pointers.

We now break CoversFor() function into two functions. The member functions of
Cover class now returns map[string]*expression.Cover instead of
map[*expression.Cover]value.Value, i.e., it now uses a string as a key to the
map such that if two expression has the same string value they'll be treated
as the same. The wrapper function CoversFor() then converts the new map with
string as key to the original map with *expression.Cover as key, so the callers
of CoversFor() do not need to be changed.

Also fixed a couple of typos in comment noticed during debugging.

Change-Id: I806cf8457ee4b90c67b661f2ad6eb869b39120c7
Reviewed-on: http://review.couchbase.org/92069
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>