MB-29587 Query with ggregation may return incorrect result if USE KEYS hint is used
When USE KEYS hint is used, the planner builds a simple KEY SCAN. When aggregation
is present, previously (in spock) we add WHERE clause as filter, unconditionally.
In Vulcan, with aggregate pushdown, we only add WHERE clause if the aggregation
has not been pushed down to index scan. The issue is that when we build KEY SCAN,
currently we only reset the pushdown of ORDER/LIMIT/OFFSET, and leaves GROUP/AGGS
pushdown still present, which leads to the WHERE clause not being added as FILTER,
and subsequently wrong result.

The fix is to reset all index pushdown related fields when building KEY SCAN.

Change-Id: I88980b5f388e566f192b3364a585bc13c09ac0bd
Reviewed-on: http://review.couchbase.org/94028
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>