MB-29743 Unnest scan does not work when UNNEST is mixed with ANSI JOIN
When UNNEST exists in a query, we infer predicate is_array(<unnest_expression>)
and add it to WHERE clause. This depends on function collectInnerUnnests(), which
in turn depends on join nodes implementing algebra.JoinTerm interface. The problem
is that algebra.AnsiJoin has Right() function returning FromTerm, not *KeyspaceTerm,
and thus it does not currently implements JoinTerm interface, as a result an AnsiJoin
node is not considered JoinTerm, which leads to problems.

The fix is to remove the function Right() from the JoinTerm interface definition.
Having a Left() and an Outer() function should be enough to identify a join node.

Change-Id: I0d2e6787d8af0b6fbff91739112055460c7df2c5
Reviewed-on: http://review.couchbase.org/94382
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>