MB-29740: concurrent map access in parsed value
The underlying issue is that join goroutine is writing value fields (while searching for t.owner) at the same time as the filter is looking for t.capped (although I would expect this to be happening at a later stage?).
The Copy() part is actually handled properly in MB-29003.
I'm adding a RWLock.

Change-Id: I51bd73065f23dde599b25bc5cf2eb8cc1a0e0d85
Reviewed-on: http://review.couchbase.org/94412
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>