MB-29709 Adding semantics layer to N1QL
N1QL currently does not have a semantics layer, as a result semantic
errors are currently either catched at the parser or the planner.
This checkin adds a semantics layer to N1QL and moved the semantic
errors added for ANSI JOIN from parser file to semantics layer. The
semantic check layer is called immediately after parsing a statement.

A few other errors, notebly ones encountered during formalization
process are also "moved" from the planner to semantics layer. Note
the move is only moving the error definition, the error number
remains the same as before (there is a special section in the new
errors/semantics.go file that includes the moved errors).

Another notable change is that for the inner keyspace of a join,
the field "keys" is currently used for both "USE KEYS" and
"ON KEYS" clauses (which are mutually exclusive), which requires
the semantic check for "USE KEYS" to be done inside the parser.
This checkin adds a new joinKeys field such that the two are
separated and thus the semantic check can happen later.

Change-Id: Ic4bd1194c0a65eb0f43c019d91395ae7d34e8b36
Reviewed-on: http://review.couchbase.org/94536
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>