MB-29710 Support expression term and subquery term as right-hand side of ANSI JOIN
Current support for ANSI JOIN in vulcan requires the right-hand side of teh join
to be a keyspace. This feature adds support for expression term and subquery term
as right-hand side of an ANSI JOIN or ANSI NEST.

A new interface, algebra.SimpleFromTerm is defined to include KeyspaceTerm,
ExpressionTerm and SubqueryTerm. Join hint and property flags are added to
ExpressionTerm and SubqueryTerm. The expression term now also includes an indicator
for whether correlation is present in the expression term. (Subquery already has
this indicator). Note that we allow FROM term expression to be correlated (e.g.,
referring to the left-hand side of the join, also called lateral correlation),
but FROM term subquery cannot be correlated.

The planner considers hash join for expression term and subquery term even if
USE HASH hint is not specified (EE only). In such cases the expression term or
the subquery term is used as build side of hash join. If hash join is not feasible,
then nested-loop join will be considered. Note that hash join does not handle
correlation and thus if an expression term is correlated, nested-loop join will
be used.

At execution time, if an expression term is not correlated, add a cache for the
results of the expression term such that if repeatedly executed we can use the
cache instead of reevaluating the expression term every time.

Change-Id: I82c9c2c8d851e3f0b9c3d2163521df1c3cefa759
Reviewed-on: http://review.couchbase.org/94692
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>