MB-29711 Support ANSI MERGE
Currently MERGE statement only supports ON KEYS clause, which is equivalent
to performing a lookup join between the source and the target of the MERGE.
This feature adds support for ANSI MERGE which performs an ANSI JOIN between
source and target of the MERGE. (The original MERGE syntax is still supported
for backward compatibility).

The target keyspace reference now supports USE INDEX hints (it is the inner
of an ANSI JOIN). The source, which can be a keyspace term, an expression
term, or a subquery term, can now have a join hint (USE HASH or USE NL).

The insert action for ANSI MERGE requires explicity document key
specification, using either (<key_expr>, <value_expr>) syntax or
(KEY <key_expr>, VALUE <value_expr>) syntax. The legacy MERGE statement
still only support <value_expr> syntax for insert action.

A new indicator for MERGE is added to indicate whether the MERGE is using
original syntax (ON KEY clause) or new syntax (ON clause).

The planner builds the MERGE as an ANSI JOIN between source and target of
the MERGE statement. At runtime the MERGE performs the lookup join if
ON KEY clause is specified (existing behavior); if ON clause is specified,
it simply gets the results from nested-loop join or hash join operator
below.

Change-Id: Id913f36241899a50fdaf757e795fdfdfb4aabae4
Reviewed-on: http://review.couchbase.org/94786
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>