MB-29712 Support ANSI OUTER JOIN to ANSI INNER JOIN transformation
ANSI OUTER JOIN to ANSI INNER JOIN transformation is a standard query
rewrite capability of relational databases. Since ANSI JOIN support
is added to Vulcan, we now add ANSI OUTER JOIN to ANSI INNER JOIN
transformation capability.

The concept is simple, for an outer join, if the document on the
dominant side is not joined with a document on the subservient
side, then the document from the dominant side is returned as a
result document; in such results the subservient side document is
MISSING. If a subsequent filter evaluation rejects such result,
e.g., when there is a filter on a field of the subservient
document, then all results where the subservient side document is
MISSING will be rejected. Such filters are called "null-rejecting"
filters in relational databases, we inherit this term although
technically it should be called "missing-rejecting" filters.

At the beginning of planning process, after we classify all WHERE
clause filters and pushable ON-clause filters, we traverse the
algebra tree looking for ANSI OUTER JOIN/NEST, and if found, look
at all filters and join filters involving the keyspace or expression
term or subquery term on the subservient side of the outer join, and
look for null-rejecting filters. If found, then we can transform the
outer join into an inner join.

Special care is needed for collection predicate such as ANY, EVERY,
ANY AND EVERY, where a binding variable can be used that references
the subservient side of the ANSI OUTER JOIN, in such cases such
binding variables are treated the same as the alias for the
subservient side of the ANSI OUTER JOIN.

In addition, UNNEST also requires special handling, since the UNNEST
can reference an array from the subservient side of an earlier ANSI
OUTER JOIN. In such cases, the alias for the UNNEST operation is
also treated the same as the alias for the subservient side of an
ANSI OUTER JOIN.

Change-Id: I1ec8cb3f881809eb7fb8d32afed688f6b516b3dc
Reviewed-on: http://review.couchbase.org/94844
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>