MB-29845 Properly handled newly created "correlated" indicator in marshalled plan for expression scan
MB-29710 introduced a correlated indicator for expression scan, in case of mixed node cluster
where a query node is running Vulcan and generate an expression scan without the correlated
indicator, and that plan is subsequently executed on a Mad-Hatter node, it may incorrectly
assume that the expression scan is not correlated and may generate wrong result if the
expression scan is indeed correlated.

The fix is the store the "uncorrelated" information in the plan, i.e., if there is nothing
stored in the plan, then we assume the expression scan is correlated. It is safe to execute
a uncorrelated expression scan as correlated (just more work), but not the reverse.

Change-Id: I599961e7197a65f15fb787d2e5607d861b88401f
Reviewed-on: http://review.couchbase.org/94934
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>