MB-29977 Correctly return correlated indicator from select term
MB-29710 adds a check for whether a subquery term in the FROM clause is
correlated. In the problematic query there is a select term in the node
tree "select ... except (select ... union select ...)" due to the
presence of parentheses after the except. Currently the code for
IsCorrelated() of a SelectTerm always returns true (with a FIXME) which
causes issues and return an error for the query.

The fix is to check the input of the SelectTerm for correlation instead
of always return true. The SetCorrelated() function is not needed (which
also has a FIXME in it) and is removed.

Change-Id: Ia087759c8cb92244a0a75d3b471f894205a67b72
Reviewed-on: http://review.couchbase.org/95261
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>