MB-25652 Use hash table for long IN-list evaluation
When an IN-list is large (number of elements in the IN-list is large), then
evaluation of the IN-list by comparing with each element in turn can be
inefficient, especially if the IN-list needs to be repeatedly evaluated.

Introduce a mechanism of using hash table for IN-list evaluation if the number
of elements in the IN-list exceeds a threshold (16), and the elements are
"static", i.e. cannot contain any references to any fields of any documents.
Host variables can be used, in particular the entire IN-list can be passed in
as a host variable, and the hash table mechansim works in such cases. A
special case is also made for an IN predicate in the form of "IN <subquery>",
and if the <subquery> is not correlated, we can also use hash table to
handled the results of the subquery (as elements in an IN-list).

If elements of an IN-list met the prerequisite, then a hash table is built
using elements of the IN-list when the IN-list is evaluated for the first
time. Subsequent evaluations of the IN-list simply probes the hash table.

The life span of the hash table is one complete cycle of exeution of an
operator. Currently this optimization is only enabled for Filter operator
(for filter expression) and ANSI JOIN/NEST operators (for ON-clause
expression).

Change-Id: I8b4a7a0864ad1c2129d32be7a49c01f51350fcf3
Reviewed-on: http://review.couchbase.org/95758
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>