MB-30262 Dynamic expansion of index spans for IN-list specified as host variables
When an IN-list is specified as a host variable, e.g., WHERE c1 IN $inlist, the
planner currently generate a single index span with ARRAY_MIN($inlist) and
ARRAY_MAX($inlist) as boundaries of the index span. This creates performance
issues in many cases.

This enhancement is to dynamically expand the index span generated by the planner
into spans with each individual element of the IN-list as boundaries, which
mimics the behavior if constants are used in the IN-list. This provides similar
performance characteristics for a dynamically specified IN-list as when the
IN-list is specified at prepare time as constants.

A threshold of 8192 spans is in place such that we don't generate too many
index spans.

Change-Id: I03c80c95815103446e45c997229783fbb494aee5
Reviewed-on: http://review.couchbase.org/96121
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>