MB-30287 ANSI JOIN converted to lookup join does not work properly if covered index scan is used for left-hand side keyspace
If the ON-clause of an ANSI JOIN is specified as meta(keyspace2).id IN <key_exprs>,
and an appropriate index cannot be found on the right-hand side keyspace, we
internally convert the ANSI JOIN to lookup join, with <key_exprs> as the ON KEYS
expression. The issue with this defect is that if the left-hand side keyspace is
using covered index scan, the covered transformation is not properly done on the
<key_expr> and thus run time evaluation of the <key_expr> do not properly return
the right result.

The fix is to perform proper covered expression transformation on the <key_expr>.

Change-Id: Ifeaedc66c62c949363164c238a10d8a4ce16a6d9
Reviewed-on: http://review.couchbase.org/96232
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: http://review.couchbase.org/96239
Tested-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Well-Formed: Build Bot <build@couchbase.com>