MB-30292 Multiple level of unnest does not properly use sargable predicate
If we have multiple levels of unnests, e.g. FROM keyspace UNNEST keyspace.arr
AS ks1 UNNEST ks1.arr2 AS ks2, and we have a join predicate that references
ks1 and ks2, currently such join predicates are not considered as index
predicates for unnest scan. We need to includes join predicates that references
any of the aliases in the multiple levels of UNNEST as long as it traces back
to the primary keyspace (the first keyspace in the WHERE clause).

Change-Id: I6bfda3ec0782b74d8824b1f8af546e174cebcf5a
Reviewed-on: http://review.couchbase.org/96286
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>