MB-30340 implement simple common table expression
I need to add tests (there aren't any tests for LET in filestore or multistore)
but this gives you an idea of how it works.
This is the sort of syntax it implements:

with myArray as (
   (select raw object_pairs(default) aaa from default use keys "k1")[0] 
) select a1.name x, a1.val y 
   from (myArray) a1 
   where a1.val within (select raw max(a2.val) from myArray a2);

I have created a parent WITH operator which sets the
target variable and then fires the underlying select.

With this implementation you can mix WITH and LET, which partially addresses
the cascaded LET issue.
I will have a look at that in a different MB.

Change-Id: Ifd0207fa576531415bd50b309ec5659c521d6783
Reviewed-on: http://review.couchbase.org/96343
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>