MB-30350 Properly handle nested field specification in Covering code
The mechanism for determining whether an expression is covered is overhauled
in Spock, and a number of defects have been fixed in this area. This defect
deals with a case where we have nested field specification in index key as
well as query predicate. For covering determination we currently have a
mechanism to "trickle" information for nested fileds but it currently only
apply to "Equivalent" case. In this particular defect we need the same
"trickle" mechansim for case of skipping (not the keyspace considered).

The fix is to rename "trickleEquiv" to "trickle" and apply the logic to both
Equivalent and Skip cases.

Change-Id: I7a0d12e6346266c66ef977cd1c66d8d74684db81
Reviewed-on: http://review.couchbase.org/96882
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>
Reviewed-on: http://review.couchbase.org/96962