MB-30495 Better identifier handling for index covering determination
For determining whether an identifier is "covered", we now rely on
information in the identifier itself rather than covering options
that indicate whether an identifier is "single" (not part of a field).
Identifiers are already labeled if it's a keyspace identifier. We
now also label identifiers as:
  - projection list alias: if an expression in the projection list
    has an alias, this alias can subsequently be used in order by
    clause currently (and eventually group by clause as well).
  - binding variable: any kind of binding, e.g., variable used in
    LET/LETTING clause, or ANY/EVERY/ANY AND EVERY clauses. E.g.,
    a LETTING variable can be used in HAVING clause.
  - unnest alias: for UNNEST operation, mark the alias used for
    UNNEST (which is marked both as a keyspace alias and an
    unnest alias)

These are the kinds of identifiers (in addition to keyspace
identifiers) that I can find so far that affects covering decisions.
As more situations surface we'll need to update both the marking of
the identifier as well as covering logic. Hopefully all identifiers
used in N1QL will be properly labeled, which should help not only
covering determination but other aspects of identifier handling as well.

Change-Id: I1037341cd8b242f79df0c82669c41d751b565a45
Reviewed-on: http://review.couchbase.org/96961
Reviewed-by: Marco Greco <marco.greco@couchbase.com>
Tested-by: Bingjie Miao <bingjie.miao@couchbase.com>