MB-30818 stop intersect scans if send receives a stop signal
For 5.5 plus execution:base.stopped is reliable and although the intersect scam will still attempt
to process the queue and send values, the send will be ignored - the fix is easy and all it does
is avoid some time wasting.

Prior to 5.5 execution:base.stopped is not reliable, and the hangs are pretty much guaranteed - the fix is much more convoluted and it cant be cherry picked.

Change-Id: I626fdc36826e5bb4b4c77dc35caab0f7db043326
Reviewed-on: http://review.couchbase.org/98013
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>