MB-27777 reuse anonymous prepared statements by sharing name
various things to note:
the feature is currently partly disabled because I need to measure the performance impact of sharing the statements
once perfrunner is run on 6.5 and get a baseline, I will enable it

planner/planner.go introduces an interface to get round the natural cyclic reference between algebra, plan, planner and prepareds
prepareds implements it.
it will come useful for the statement cache too

currently we store in the prepared cache the whole statement, which is ugly and gives rise to a whole bunch of
undesirable behaviours.
when I changed that for explain we had two whole months of fixing testrunner tests, so, as desirable as it may be,
I hacked my way around to implement the 'force' keyword.
We will deal with it properly in another MB (to be logged) if we think it's required.

Change-Id: Ib07ece7feb4fe75126afd980722ce87659c88b6f
Reviewed-on: http://review.couchbase.org/98270
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>