MB-27778 deprecate encoded_plan
The SDKs are not ready to let go of encoded_plan as of yet, so with this change
we are disabling encoded_plan as long as the prepared is found one way or another.
For homogenous clusters, this means that the statement will be found by either push or pull,
and encoded_plan will be ignored anyway.
This is not a completed fix, but it's start.

Change-Id: I575f556bb019461f8838dfc2d5baf6f81f047591
Reviewed-on: http://review.couchbase.org/98482
Reviewed-by: Johan Larson <johan.larson@couchbase.com>
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>