MB-30924 improve REST API parsing
Have streamlined newHttpRequest and form parsing (eg searching for individual parameters - and we check all one by one!)
involved ranging all passed parameters, same for named and positional args, searching for a prepared statement involved
delayed json parsing a couple of times!)
Have added a check for statement + prepared that was missing, and added an extra test.

This shaves a few percentage points in ServeHttp, but is not enough.
Now the worse offender is util:UUID() which spends lots of time in crypto/rand. 

Change-Id: I9be898b330ace9045a2c4f528308f363f417f30b
Reviewed-on: http://review.couchbase.org/98594
Reviewed-by: Sitaram Vemulapalli <sitaram.vemulapalli@couchbase.com>
Reviewed-by: Bingjie Miao <bingjie.miao@couchbase.com>
Tested-by: Marco Greco <marco.greco@couchbase.com>